requirements.txt
# Core RAG Framework
langchain>=0.1.0
langchain-community>=0.0.10
langchain-groq>=0.1.0

# Vector Database
chromadb>=0.4.15

# Embeddings (Free, local, and API-based options)
sentence-transformers>=2.2.2
huggingface-hub>=0.19.0

# Security & Environment
python-dotenv>=1.0.0

# Data Processing
pypdf>=3.17.0
python-multipart>=0.0.6

# Testing
pytest>=7.4.0

# Utilities
pydantic>=2.5.0
tqdm>=4.66.0
